+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @echojs-ecosystem/core
 
+## 0.4.0
+
+### Patch Changes
+
+- @echojs-ecosystem/reactivity@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes