+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @echojs-ecosystem/devtools
 
+## 0.4.0
+
 ## 0.3.0
 
 ### Minor Changes