+++ packages/router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @echojs-ecosystem/router
 
+## 0.4.0
+
+### Patch Changes
+
+- @echojs-ecosystem/reactivity@0.4.0
+- @echojs-ecosystem/hyperdom@0.4.0
+- @echojs-ecosystem/url-state@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes